requirements.txt
Babel==2.7.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
fuzzywuzzy==0.17.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mongoengine==0.18.2
Pillow==6.2.0
pycparser==2.19
pymongo==3.9.0
pytz==2019.3
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1
